+++ api/transactions/transactions.go
@@ -1,0 +1,7 @@
+package transactions
+
+func GetTransactions() int8 {
+	var transactionAmount int8 = 10
+
+	return transactionAmount
+}
