+++ go.mod
@@ -1,9 +1,7 @@
-module spend_tracker_go
-
-go 1.17
+module spend_track_go
 
 require (
-	github.com/gin-contrib/sse v0.1.0 // indirect
+  github.com/gin-contrib/sse v0.1.0 // indirect
 	github.com/gin-gonic/gin v1.7.7 // indirect
 	github.com/go-playground/locales v0.14.0 // indirect
 	github.com/go-playground/universal-translator v0.18.0 // indirect
@@ -24,3 +22,5 @@ require (
 	google.golang.org/protobuf v1.27.1 // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
+
+go 1.17
