--- pkg/api/transactions/transactions.go
@@ -1,7 +1,0 @@
-package transactions
-
-func GetTransactions() int8 {
-	var transactionAmount int8 = 10
-
-	return transactionAmount
-}
